requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
torchaudio
transformers
datasets
beautifulsoup4
requests
scrapy
matplotlib
seaborn
scipy
streamlit
feedparser
langdetect
fasttext-wheel
